requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
bcrypt==3.2.2
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.78.0
fastapi-mail==1.0.9
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==3.0.0
packaging==21.3
passlib==1.7.4
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-multipart==0.0.5
redis==4.3.3
requests==2.28.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.1
yarg==0.1.9
